# public/openapi.yaml
openapi: 3.0.0
info:
  title: The R-hub R versions API
  version: 1.0.0
  contact:
    email: admin@r-hub.io
  license:
    name: The MIT License
    url: 'https://opensource.org/licenses/MIT'
  x-logo:
    url: 'https://github.com/r-hub/branding/raw/master/logo/rhub-header.png'
  description: |
    This service is an API to query current and past R versions
    and installers. You can use it is devops scripts, e.g. for continuous
    integration tools, to query which version(s) or R you want to install.
    It is updated every hour, so a new R release or a new installer is
    typically picked up within on hour.

    ### Format

    This is a JSON API, all responses are JSON.

    ### The R release process

    In the current R release process, there is a new major or minor release
    every year, typically in April. There are also several patch releases.
    Once there is a newer major or minor release, the older major or minor
    releases typically do not get new patch releases. (E.g. once R 3.6.0 is out,
    the R 3.5.x minor branch will not receive more releases.)

    ### Typical use cases

    * To report the currently released version of R, use `/r-release`.
    * To download and compile the currently released version of R, use
      `/r-release-tarball`.
    * To download the latest Windows or macOS installer, use `/r-release-win`
      or `/r-release-macos`, respectively.
    * To report or download the source or an installer for the latest release
...

...

...

...

...

...

...
servers:
...
paths:
...
  '/r-oldrel/:n':
    parameters:
    - name: n
      in: path
      required: true
      description: Number of minor versions to go back.
      schema:
        type: integer
    get:
      summary: Legacy R versions
      description: |
        Last releases of previous minor branches. E.g. `r-oldrel/2` is the
        release that was `oldrel` before the current R release. `r-oldrel/1`
        is the same as `r-oldrel`.
      responses:
        '200':
          description: Successful query.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/version'
...
components:
  schemas:
    version:
      properties:
        version:
          type: string
          pattern: '^\d+\.\d+(\.\d+)?$'
        date:
          type: string
          format: date-time
        nickname:
          type: string
      required:
        - version
        - date
        - nickname
      example:
        version: '3.6.3'
        date: "2020-02-29T08:05:16.744223Z"
        nickname: "Holding the Windsock"
...